Rename adapter
